move scm payload classes out of model/entity namespace
they don't belong there, not sure why added there in first place
